Import fetch_dataset from lisbet.hub in tests

+++ src/lisbet/training/core.py
@@ -147,8 +147,11 @@ def _configure_optimizer_and_scheduler(model, learning_rate):
 def _configure_dataloaders(tasks, group, batch_size, sample_ratio, pin_memory):
     """Internal helper. Configures dataloaders for a group."""
     # Estimate number of samples
-    n_batches = np.ceil(
-        min(getattr(task, f"{group}_dataset").n_frames for task in tasks) / batch_size
+    n_batches = int(
+        np.ceil(
+            min(getattr(task, f"{group}_dataset").n_frames for task in tasks)
+            / batch_size
+        )
     )
     if sample_ratio is not None:
         n_batches = int(n_batches * sample_ratio)
